backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
motor==3.6.0
beanie==1.27.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.0
email-validator==2.2.0
pytesseract==0.3.13
Pillow==11.0.0
PyMuPDF==1.25.1
datasets>=2.18.0,<4.0.0
huggingface_hub>=0.21.0
textstat>=0.7.3
scikit-learn>=1.4.0
evaluate>=0.4.0
rouge-score>=0.1.2
sacrebleu>=2.3.1
seqeval>=1.2.2
accelerate>=0.26.0
torch>=2.1.0
